Add GUI, as well as added duke package and GUI package
