+++ data/duke.txt
@@ -1,5 +1,4 @@
-[T][ ] a
-[T][ ] b
 [T][ ] cry
 [D][ ] i want to cry (by: 1998-10-10 10:10)
 [E][ ] crying (from: 1998-10-10 10:10 to: 2000-11-12 11:11)
+[T][ ] A
