+++ src/test/java/duke/task/ToDoTest.java
@@ -1,0 +1,17 @@
+package duke.task;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import duke.task.Task;
+import duke.task.ToDo;
+import org.junit.jupiter.api.Test;
+
+class ToDoTest {
+    @Test
+    public void testToString() {
+        Task task = new ToDo("eat");
+        String actualOutput = task.toString();
+        String expectedOutput = "[T][ ] eat";
+        assertEquals(expectedOutput, actualOutput);
+    }
+}
